requirement.txt
coremltools
torchvision 
cntk==2.6
tf2onnx
mxnet
chainer
onnx-caffe2
git+https://github.com/asiryan/caffe2onnx
onnx2keras
onnx-pytorch
onnx-coreml
onnx-tf
onnxmltools
